backend/etl/requirements.txt
kafka-python
requests
python-dotenv
pytest
